Expose last option added via last().
So that option and command builders can modify the last option added.

This is necessary as option() and command() return the program and some
builder logic needs to modify the option or command after addition, for
example the command module defines options and commands from markdown
documents and must merge other data such as converter() with the last
option added.